Move active task transient
